1-A

+++ src/twoDShapes/Closed2DShapes.java
@@ -1,6 +1,6 @@
 package twoDShapes;
 
 public abstract class Closed2DShapes{
-	abstract int surfaceArea();
-	abstract int perimeter();
+	abstract float surfaceArea();
+	abstract float perimeter();
 }
